requirements.txt
Babel==1.3
Django==1.8.13
MySQL-python==1.2.5
Pillow==2.7.0
django-compressor==1.4
django-csvimport==2.4
django-extra-views==0.6.4
django-grappelli==2.6.3
django-haystack==2.3.1
django-localflavor==1.1
django_markdown==0.8.4
django-nested-inline==0.3.6
django-oscar==1.1.1
django-oscar-accounts==0.3
django-oscar-datacash==0.8.3
django-oscar-paypal==0.9.5
django-treebeard==3.0
django-widget-tweaks==1.3
html5lib==0.999
paypal==1.2.1
pycountry==1.10
python-amazon-mws==0.6
xlrd==0.9.2
lxml==3.4.2
beautifulsoup4==4.3.2
pycurl==7.19.3.1
